cleanup: move FilterBar into shared components
